+++ appdaemontestframework/given_that.py
@@ -1,3 +1,5 @@
+from datetime import datetime
+
 from appdaemontestframework.common import AppdaemonTestFrameworkError
 
 
@@ -46,7 +48,17 @@ class GivenThatWrapper:
                 if attribute is None:
                     return state['main']
                 elif attribute == 'all':
-                    return state['attributes']
+                    def format_time(timestamp: datetime):
+                        if not timestamp: return None
+                        return timestamp.isoformat()
+
+                    return {
+                        "last_updated": format_time(state['last_updated']),
+                        "last_changed": format_time(state['last_changed']),
+                        "state": state["main"],
+                        "attributes": state['attributes'],
+                        "entity_id": entity_id,
+                    }
                 else:
                     return state['attributes'].get(attribute)
 
@@ -68,12 +80,18 @@ class GivenThatWrapper:
         given_that_wrapper = self
 
         class IsWrapper:
-            def is_set_to(self, state, attributes=None):
+            def is_set_to(self,
+                          state,
+                          attributes=None,
+                          last_updated: datetime = None,
+                          last_changed: datetime = None):
                 if not attributes:
                     attributes = {}
                 given_that_wrapper.mocked_states[entity_id] = {
                     'main': state,
-                    'attributes': attributes
+                    'attributes': attributes,
+                    'last_updated': last_updated,
+                    'last_changed': last_changed
                 }
 
         return IsWrapper()